python_client/requirements.txt
#!/usr/bin/env python3
"""
Requirements file for the Python WebSocket client simulator
"""

websocket-client==1.6.4
